LayoutTests/storage/indexeddb/setVersion-null-expected.txt

Test IndexedDB with null version string

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
PASS Boolean(indexedDB && IDBCursor && IDBDatabase && IDBDatabaseError && IDBDatabaseException && IDBFactory && IDBIndex && IDBKeyRange && IDBObjectStore && IDBRequest && IDBTransaction) is true

indexedDB.open(name, description)
db = event.target.result
db.setVersion(null);
PASS db.version is 'null'
PASS successfullyParsed is true

TEST COMPLETE
